add threads API
