+++ src/controllers/index.ts
@@ -1,4 +1,4 @@
 export { A2AController } from "./a2a.controller.js";
 export { ModelApiController } from "./api/model.api.controller.js";
-export { ThreadApiController } from "./api/thread.api.controller.js";
+export { ThreadApiController } from "./api/threads.api.controller.js";
 export { QueryController } from "./query.controller.js";
