+++ src/routes/index.ts
@@ -1,5 +1,5 @@
 export { createA2ARouter } from "./a2a.routes.js";
 export { createModelApiRouter } from "./api/model.routes.js";
-export { createThreadApiRouter } from "./api/thread.routes.js";
+export { createThreadApiRouter } from "./api/threads.routes.js";
 export { createApiRouter } from "./api.routes.js";
 export { createQueryRouter } from "./query.routes.js";
